requirements.txt
PyYAML>=5.4
pyyaml~=6.0
seaborn~=0.11.2
joblib>=1.2.0
scikit-learn~=1.0.1
numpy>=1.21.2
pandas
matplotlib
gradio
